bpf: add kfunc_md for tracing-multi
Introduce the kfunc_md, which is used to store the BPF progs and the
function metadata for tracing multi-link. rhashtable is used here. The key
of it is the kernel function address, and the value is the pointer to
struct kfunc_md. Following data is stored in the struct kfunc_md:

- The BPF progs, whose type is FENTRY, FEXIT or MODIFY_RETURN. The struct
  kfunc_md_tramp_prog is introduced to store the BPF prog and the cookie.
- The kernel function address
- The kernel function arguments count
- If origin call needed

Link: https://lore.kernel.org/bpf/CADxym3anLzM6cAkn_z71GDd_VeKiqqk1ts=xuiP7pr4PO6USPA@mail.gmail.com/
Link: https://lore.kernel.org/bpf/CAADnVQ+G+mQPJ+O1Oc9+UW=J17CGNC5B=usCmUDxBA-ze+gZGw@mail.gmail.com/
Signed-off-by: Menglong Dong <dongml2@chinatelecom.cn>
---
v3:
- implement the function metadata with rhashtable

v2:
- implement the function metadata with hash table, as Alexei advised