x86,bpf: add bpf_global_caller for global trampoline
Implement the bpf global trampoline "bpf_global_caller" for x86_64. Thanks
to Alexei's advice, we implement most of the global trampoline with C
instead of asm.

We implement the entry of the trampoline with a "__naked" function, who
will save the regs to an array on the stack and call
bpf_global_caller_run(). The entry will pass the address of the array and
the address of the rip to bpf_global_caller_run().

In bpf_global_caller_run(), we will find the metadata by the function ip.
For origin call case, we call percpu_ref_get() before call the origin
function to protect the metadata, which is similar to __bpf_tramp_enter().
Then we will call all the BPF progs, just like what BPF trampoline do.
Sleepable bpf prog is not supported(for now), so we can run all the progs
within rcu_read_lock_dont_migrate().

Without origin call, the bpf_global_caller_run() will return NULL. Then,
the entry will restore the regs and return; In origin call case, it will
return the address where we store the return value, and the entry will
load the return value to rax and make the RSP skip the rip before return.

Signed-off-by: Menglong Dong <dongml2@chinatelecom.cn>
---
v3:
- don't support sleepable prog
- use the rcu_read_lock_dont_migrate() globally
- use a single global caller, instead of creating 7 instance for different
  function argument count

v2:
- rewrite the global trampoline with C instead of asm