bpf: support mixed use of tracing and tracing-multi
For now, the bpf global trampoline can't work together with trampoline.
For example, we will fail on attaching the FENTRY_MULTI to the functions
that FENTRY exists, and FENTRY will also fail if FENTRY_MULTI exists.

We make the global trampoline work together with trampoline in this
commit.

tracing-multi mix tracing
-------------------------
We introduce the function bpf_gtrampoline_link_shim() to make
bpf_gtramp_link adjust bpf_trampoline, which will be called in
bpf_gtrampoline_link_prog(). If the bpf_trampoline of the function exist
in the kfunc_md or we find it with bpf_trampoline_lookup_exist(), it means
that we need do the adjusting. It's simple, we create a
bpf_shim_tramp_link for the prog and link it to the trampoline with
bpf_trampoline_link_gtramp_shim().

tracing mix tracing-multi
-------------------------
When attaching a trampoline, we will check if a global trampoline is
already attached on the target function. If so, we find all the bpf progs
from the kfunc_md and link them to the trampoline with
bpf_trampoline_link_gtramp_shim().

In __bpf_gtrampoline_unlink_prog(), we will call
bpf_trampoline_unlink_gtramp_shim() to release the bpf_shim_tramp_link,
and the bpf prog will be unlinked in bpf_link_free().

Signed-off-by: Menglong Dong <dongml2@chinatelecom.cn>