ftrace: supporting replace direct ftrace_ops
Introduce the function replace_ftrace_direct(). This is used to replace
the direct ftrace_ops for a function, and will be used in the next patch.

For example, the function "foo" is hooked by the direct ftrace_ops "A".
Now, we want to hook the function "foo" with the direct ftrace_ops "B",
which means we need replace the ftrace_ops "A" with "B". The function that
we introduce is for this propose.

In fact, the ftrace_ops "A" means the global trampoline, and the
ftrace_ops "B" means the bpf trampoline. When the FENTRY prog tries to
hook a kernel function that is already hooked by the FENTRY_MULTI(global
trampoline), we can replace the global trampoline with the bpf trampoline
by using the "replace_ftrace_direct".

To achieve this goal, first, we register B directly, and the callback of
the functions in A and B will fallback to the ftrace_ops_list case.

Then, we modify the address of the entry in the direct_functions to
B->direct_call, and remove it from A. This will update the dyn_rec and
make the functions call b->direct_call directly. If no function in
A->filter_hash, just unregister it.

So a record can have more than one direct ftrace_ops, and we need check
if there is any direct ops for the record before remove the
FTRACE_OPS_FL_DIRECT in __ftrace_hash_rec_update().

Signed-off-by: Menglong Dong <dongml2@chinatelecom.cn>