bpf: tracing: add multi-link support
In this commit, we add the support to allow attaching a tracing BPF
program to multi hooks, which is similar to BPF_TRACE_KPROBE_MULTI.

The use case is obvious. For now, we have to create a BPF program for each
kernel function, for which we want to trace, even through all the program
have the same (or similar logic). This can consume extra memory, and make
the program loading slow if we have plenty of kernel function to trace.
The KPROBE_MULTI maybe a alternative, but it can't do what TRACING do. For
example, the kretprobe can't obtain the function args, but the FEXIT can.

For now, we support to create multi-link for fentry/fexit/modify_return
with the following new attach types that we introduce:

  BPF_TRACE_FENTRY_MULTI
  BPF_TRACE_FEXIT_MULTI
  BPF_MODIFY_RETURN_MULTI

The bpf prog is not BTF type aware during loading, which means we can't
access the function argument directly with ctx[x]. Instead, we need use
the bpf_get_func_arg() to obtain the function argument, and use
bpf_core_cast() to do the casting.

Signed-off-by: Menglong Dong <dongml2@chinatelecom.cn>
---
v3:
- don't record the accessed function argument and don't do the argument
  consistency checking during attaching. Using the bpf_get_func_arg() to
  obtain the function argument instead.